Cleanup some code and make so multiple consumers can work before I break it
